requirements.txt
flask
pandas
pvlib
psutil
python-dotenv
requests
pymongo
